Add ScaledFileImage class
Add a new ImageProvider class, which is closely modeled after FileImage,
but supports specifying a targetHeight and targetWidth. This is handed
over to the Codec, which scales the image after decoding. This is very
useful to avoid high memory usage and cpu/gpu load for repeated sclaing,
when showing high resolution images on small widgets.

Signed-off-by: Julian Scheel <julian@jusst.de>