CachedNetworkImage: Prescale images to match size
If a widget size is specified, pass this to the ScaledImageFile, so that
the image is prescaled right after decoding. This way the memory usage
is reduced when working with high resolution images which are only
rendered to small targets. Furthermore the load while rendering is
reduced as the images are cached in the exact target size.

Signed-off-by: Julian Scheel <julian@jusst.de>